refactor(web): extract param-control components from App.tsx (issue #642, Phase 3 seam 2 pt 2)
Moves the prop-driven param-control components out of App.tsx verbatim:

- src/components/params/Knob.tsx — the rotary knob (param + vfo variants)
- src/components/params/GeometryCombobox.tsx — searchable antenna picker
- src/components/params/BandDropdown.tsx — click-only band popover
- src/components/params/ParamForm.tsx — schema-driven field renderer,
  plus its private layoutStyle() helper
- src/components/AwaitingTrustPanel.tsx — the untrusted-design gate panel

DesignLoadError/DesignAdvisory and KnobOpt stay in App.tsx (both still read
by the main component's own state/hooks) and are now exported so
AwaitingTrustPanel and ParamForm can import them by type only — an
erased, no-runtime-cycle import, required anyway by the project's
isolatedModules setting.

App.tsx drops its now-unused CSSProperties and mixHex imports (both were
only reachable through the moved layoutStyle/Knob code).
