+++ src/antennaknobs/web/frontend/src/App.tsx
@@ -9,7 +9,6 @@ import {
   useState,
   useSyncExternalStore,
 } from "react";
-import type { CSSProperties } from "react";
 import {
   backendAllowed,
   BACKEND_LABEL,
@@ -46,9 +45,8 @@ import {
   formatSwr,
   reflectionCoefficient,
 } from "./lib/format";
-import { mixHex, richardsonExtrap } from "./lib/math";
+import { richardsonExtrap } from "./lib/math";
 import {
-  applyVisibility,
   familyOf,
   familyRank,
   FAMILY_LABELS,
@@ -59,7 +57,6 @@ import {
   snapForExample,
   type BandSpec,
   type ExampleDescriptor,
-  type KnobLayout,
   type ParamValueBag,
   type ResultFieldSpec,
   type ResultSchemaItem,
@@ -79,6 +76,11 @@ import {
 } from "./lib/view";
 import { BackendConfigModal } from "./components/backend/BackendConfigModal";
 import { KnobMenuNumber, NumberField } from "./components/backend/fields";
+import { AwaitingTrustPanel } from "./components/AwaitingTrustPanel";
+import { BandDropdown } from "./components/params/BandDropdown";
+import { GeometryCombobox } from "./components/params/GeometryCombobox";
+import { Knob } from "./components/params/Knob";
+import { ParamForm } from "./components/params/ParamForm";
 
 type Wire = {
   label: string;
@@ -94,869 +96,18 @@ type Wire = {
 
 // One advisory finding from the design screener (what a design does that a
 // typical one doesn't), attached to a trust-required entry.
-type DesignAdvisory = { severity: string; message: string; line: number };
+export type DesignAdvisory = { severity: string; message: string; line: number };
 
 // A user design reported by GET /examples that didn't register. Either a real
 // load error (bad Python), or — when `trust_required` — a design that loaded
 // fine but hasn't been trusted to run yet, carrying its screener `advisory`.
-type DesignLoadError = {
+export type DesignLoadError = {
   name: string;
   file: string;
   message: string;
   trust_required?: boolean;
   advisory?: DesignAdvisory[];
 };
-
-// A dependency-free rotary knob — a drop-in alternative to the range
-// slider for float/int params. Semantically a slider (role="slider"), so
-// it stays keyboard- and screen-reader-accessible: vertical drag, scroll
-// wheel, and arrow keys all adjust the value; double-click (or Enter) to
-// type an exact number. The dial sweeps ~270° from min (lower-left) to
-// max (lower-right). Absolute-angle dragging is deliberately avoided —
-// drag is a *relative* vertical delta, which is far easier to do
-// precisely than chasing the pointer around a circle.
-function Knob({
-  knobId,
-  value,
-  min,
-  max,
-  step,
-  precision,
-  unit,
-  label,
-  onChange,
-  startDeg = -135,
-  sweepDeg = 270,
-  variant = "param",
-  disabled = false,
-}: {
-  // Stable id, emitted as data-knob-id for testing/debugging.
-  knobId: string;
-  value: number;
-  min: number;
-  max: number;
-  step: number;
-  precision: number;
-  unit: string | null;
-  label: string;
-  onChange: (v: number) => void;
-  // Dial geometry in clock-angle degrees (0 = 12 o'clock, +CW). The default is
-  // the classic 270° gauge sweeping clockwise from 7:30 (lower-left) to 4:30.
-  // Pass startDeg=90, sweepDeg=-(max-min) for a CCW dial starting at 3 o'clock
-  // (elevation: -90 quarter-arc; azimuth: -360 full circle) — degrees then map
-  // 1:1 onto the dial face.
-  startDeg?: number;
-  sweepDeg?: number;
-  // "vfo" = the big weighted measurement-freq tuning dial: knurled skirt +
-  // finger dimple on an eased rotor, outer band arc that warms toward the edge.
-  // "param" = the compact setup/cut dials (clean accent, no warming).
-  variant?: "param" | "vfo";
-  // Locked (e.g. measurement freq while "lock to design freq" is on): dims the
-  // dial and ignores drag/wheel/keys.
-  disabled?: boolean;
-}) {
-  const wrapRef = useRef<HTMLDivElement>(null);
-  const dragRef = useRef<{ y: number; v: number } | null>(null);
-  const [editing, setEditing] = useState(false);
-  const [dragging, setDragging] = useState(false);
-  const isVfo = variant === "vfo";
-  const span = max - min || 1;
-  const clamp = (v: number) => Math.min(max, Math.max(min, v));
-  // Clamp + round to the param's precision so we emit clean values (2.46, not
-  // 2.4600000001) and don't spam the live solve with fp noise. No grid snap —
-  // used where the exact value matters (Home/End reaching min/max).
-  const roundP = (v: number) => {
-    const p = precision >= 0 ? precision : 6;
-    return clamp(Number(v.toFixed(p)));
-  };
-  // Snap to the nearest multiple of `step` — a clean grid anchored at 0, not at
-  // `min`. Anchoring at min offset the whole grid by min's fractional part, so
-  // nudging an off-grid value kept that offset (1.03 + 0.2 -> 1.23). Anchored at
-  // 0 it lands on a round increment (1.03 + 0.2 -> 1.2). min/max are bounds, not
-  // the grid origin; they stay reachable exactly via roundP (Home/End).
-  const snap = (v: number) => {
-    if (step > 0) v = Math.round(v / step) * step;
-    return roundP(v);
-  };
-
-  const frac = Math.min(1, Math.max(0, (value - min) / span));
-  const ang = startDeg + frac * sweepDeg;
-  const Rarc = isVfo ? 42 : 38;
-  const polar = (deg: number, r: number): [number, number] => {
-    const a = (deg * Math.PI) / 180;
-    return [50 + r * Math.sin(a), 50 - r * Math.cos(a)];
-  };
-  const arc = (r: number, a0: number, a1: number): string => {
-    const [x0, y0] = polar(a0, r);
-    const [x1, y1] = polar(a1, r);
-    const delta = a1 - a0;
-    const large = Math.abs(delta) > 180 ? 1 : 0;
-    // Sweep-flag follows the traversal direction: clockwise (SVG +angle) for an
-    // increasing clock-angle, counter-clockwise for a decreasing one. Lets a
-    // negative sweepDeg (CCW dials: elevation, azimuth) bend the correct way.
-    const sweep = delta >= 0 ? 1 : 0;
-    return `M ${x0.toFixed(2)} ${y0.toFixed(2)} A ${r} ${r} 0 ${large} ${sweep} ${x1.toFixed(2)} ${y1.toFixed(2)}`;
-  };
-  // Param-knob indicator notch across the cap face (center-out), at value.
-  const [nx0, ny0] = polar(ang, 3);
-  const [nx1, ny1] = polar(ang, 14);
-  // Only the VFO's band arc warms --accent -> --hot over the top ~40% of travel
-  // ("redlining" near the band edge). Small knobs stay a clean accent.
-  const warm = Math.max(0, (frac - 0.6) / 0.4);
-  const fillColor = isVfo ? mixHex("#2f5fb0", "#cf7a22", warm) : undefined;
-
-  // Scroll wheel: ±step per detent (×10 with Shift). Attached natively so
-  // we can preventDefault — React's onWheel is passive and can't stop the
-  // page from scrolling under the dial.
-  useEffect(() => {
-    const el = wrapRef.current;
-    if (!el) return;
-    const onWheel = (e: WheelEvent) => {
-      if (disabled) return;
-      e.preventDefault();
-      const dir = e.deltaY < 0 ? 1 : -1;
-      const mult = e.shiftKey ? 10 : 1;
-      let v = value + dir * step * mult;
-      // Snap to the step grid anchored at 0 (clean multiples), matching snap().
-      if (step > 0) v = Math.round(v / step) * step;
-      const p = precision >= 0 ? precision : 6;
-      onChange(Math.min(max, Math.max(min, Number(v.toFixed(p)))));
-    };
-    el.addEventListener("wheel", onWheel, { passive: false });
-    return () => el.removeEventListener("wheel", onWheel);
-  }, [value, step, min, max, precision, onChange, disabled]);
-
-  const onPointerDown = (e: React.PointerEvent) => {
-    if (editing || disabled) return;
-    (e.target as Element).setPointerCapture(e.pointerId);
-    dragRef.current = { y: e.clientY, v: value };
-    setDragging(true);
-    wrapRef.current?.focus();
-    e.preventDefault();
-  };
-  const onPointerMove = (e: React.PointerEvent) => {
-    const d = dragRef.current;
-    if (!d) return;
-    const dy = d.y - e.clientY; // drag up = increase
-    const sens = e.shiftKey ? 0.25 : 1; // hold Shift for fine control
-    onChange(snap(d.v + (dy / 180) * span * sens)); // ~180px = full sweep
-  };
-  const endDrag = (e: React.PointerEvent) => {
-    dragRef.current = null;
-    setDragging(false);
-    (e.target as Element).releasePointerCapture?.(e.pointerId);
-  };
-
-  // Apply a single nav/edit key to this knob. Shared by the local onKeyDown
-  // (when the knob is focused) and the global sticky-selection router (when it
-  // isn't). Returns true when the key was consumed, so callers can
-  // preventDefault only for keys we actually handled.
-  const applyKey = (key: string): boolean => {
-    if (disabled) return false;
-    if (key === "Enter") {
-      setEditing(true);
-      return true;
-    }
-    let next: number | null = null;
-    switch (key) {
-      case "ArrowUp":
-      case "ArrowRight":
-        next = value + step;
-        break;
-      case "ArrowDown":
-      case "ArrowLeft":
-        next = value - step;
-        break;
-      case "PageUp":
-        next = value + step * 10;
-        break;
-      case "PageDown":
-        next = value - step * 10;
-        break;
-      case "Home":
-        // Jump to exactly min/max — these are bounds, not grid points, so skip
-        // the step snap (roundP clamps + rounds to precision only).
-        onChange(roundP(min));
-        return true;
-      case "End":
-        onChange(roundP(max));
-        return true;
-      default:
-        return false;
-    }
-    onChange(snap(next));
-    return true;
-  };
-  const onKeyDown = (e: React.KeyboardEvent) => {
-    if (applyKey(e.key)) e.preventDefault();
-  };
-
-  const commit = (raw: string) => {
-    const n = Number(raw);
-    if (Number.isFinite(n)) onChange(snap(n));
-    setEditing(false);
-  };
-
-  const p = Math.max(0, precision);
-  return (
-    <div
-      className={`knob${isVfo ? " is-vfo" : ""}${disabled ? " is-disabled" : ""}`}
-      data-knob-id={knobId}
-      ref={wrapRef}
-      role="slider"
-      tabIndex={editing || disabled ? -1 : 0}
-      aria-label={label}
-      aria-valuemin={min}
-      aria-valuemax={max}
-      aria-valuenow={value}
-      aria-valuetext={`${value.toFixed(p)}${unit ?? ""}`}
-      aria-disabled={disabled || undefined}
-      onKeyDown={onKeyDown}
-    >
-      {editing ? (
-        <input
-          className="knob-edit"
-          type="number"
-          autoFocus
-          defaultValue={value}
-          min={min}
-          max={max}
-          step={step}
-          onBlur={(e) => commit((e.target as HTMLInputElement).value)}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") commit((e.target as HTMLInputElement).value);
-            else if (e.key === "Escape") setEditing(false);
-            e.stopPropagation();
-          }}
-        />
-      ) : (
-        <svg
-          // The VFO's 270° gauge opens at the bottom (~6 o'clock) and its
-          // content stops by y≈82, so crop the empty bottom off the box rather
-          // than reserve a full square. NOT for the others: the azimuth dial is
-          // a full circle that uses the bottom.
-          viewBox={isVfo ? "0 0 100 88" : "0 0 100 100"}
-          onPointerDown={onPointerDown}
-          onPointerMove={onPointerMove}
-          onPointerUp={endDrag}
-          onPointerCancel={endDrag}
-          onDoubleClick={() => setEditing(true)}
-        >
-          <path
-            className="knob-track"
-            d={arc(Rarc, startDeg, startDeg + sweepDeg)}
-          />
-          <path
-            className="knob-fill"
-            style={fillColor ? { stroke: fillColor } : undefined}
-            d={arc(Rarc, startDeg, ang)}
-          />
-          {/* Focus ring drawn in SVG space so it's a true circle centered on the
-              dial (50,50) — the wrapper's box-shadow ring would be an ellipse on
-              the VFO's non-square 112×99 box. Just outside the gauge arc (Rarc).
-              Shown on keyboard focus via .vfo-focus-ring CSS. */}
-          {isVfo && <circle className="vfo-focus-ring" cx="50" cy="50" r="50.5" />}
-          {isVfo ? (
-            // The whole knob body spins; the band arc behind it stays put — a
-            // fixed scale with a turning dial, exactly like a transceiver VFO.
-            <g
-              className={`knob-rotor${dragging ? " no-ease" : ""}`}
-              style={{ transform: `rotate(${ang.toFixed(2)}deg)` }}
-            >
-              <circle className="knob-cap" cx="50" cy="50" r="30" />
-              {Array.from({ length: 30 }, (_, i) => {
-                const a = (i * 360) / 30;
-                const [sx0, sy0] = polar(a, 27.5);
-                const [sx1, sy1] = polar(a, 32);
-                return (
-                  <line
-                    key={i}
-                    className="knob-skirt"
-                    x1={sx0.toFixed(2)}
-                    y1={sy0.toFixed(2)}
-                    x2={sx1.toFixed(2)}
-                    y2={sy1.toFixed(2)}
-                  />
-                );
-              })}
-              <line className="knob-notch" x1="50" y1="39" x2="50" y2="24" />
-              <circle className="knob-dimple" cx="50" cy="30" r="3.4" />
-            </g>
-          ) : (
-            <>
-              <circle className="knob-cap" cx="50" cy="50" r="15" />
-              <line
-                className="knob-notch"
-                x1={nx0.toFixed(2)}
-                y1={ny0.toFixed(2)}
-                x2={nx1.toFixed(2)}
-                y2={ny1.toFixed(2)}
-              />
-            </>
-          )}
-        </svg>
-      )}
-    </div>
-  );
-}
-
-// Designs that loaded clean but haven't been trusted to run yet. A user design
-// is a Python program that runs with your privileges, so it executes only once
-// you trust it (see design_trust.py). This panel is collapsed to one line by
-// default — click a design to see what it does (the screener advisory) and to
-// trust it. Enforcement happens at scan time (untrusted files are never
-// executed); this is just where you make the decision, per design.
-function AwaitingTrustPanel({
-  designs,
-  busy,
-  onTrust,
-}: {
-  designs: DesignLoadError[];
-  busy: string | null;
-  onTrust: (stem: string, allowEdits: boolean) => void;
-}) {
-  const [expanded, setExpanded] = useState(false);
-  const [openName, setOpenName] = useState<string | null>(null);
-  const n = designs.length;
-  return (
-    <div className="design-trust-panel">
-      <button
-        className="design-trust-summary"
-        aria-expanded={expanded}
-        onClick={() => setExpanded((v) => !v)}
-      >
-        <span className="design-trust-lock" aria-hidden="true">
-          🔒
-        </span>
-        {n} design{n === 1 ? " needs" : "s need"} your OK to run
-        <span className="design-trust-caret" aria-hidden="true">
-          {expanded ? "▾" : "▸"}
-        </span>
-      </button>
-      {expanded && (
-        <ul className="design-trust-list">
-          {designs.map((d) => {
-            const open = openName === d.name;
-            const isBusy = busy === d.name;
-            return (
-              <li key={d.name} className="design-trust-item">
-                <button
-                  className="design-trust-item-head"
-                  aria-expanded={open}
-                  onClick={() => setOpenName(open ? null : d.name)}
-                >
-                  <code>{d.name}</code>
-                  <span className="design-trust-caret" aria-hidden="true">
-                    {open ? "▾" : "▸"}
-                  </span>
-                </button>
-                {open && (
-                  <div className="design-trust-detail">
-                    {d.advisory && d.advisory.length > 0 ? (
-                      <>
-                        <div className="design-trust-advisory-head">
-                          Heads up — this design does things a normal antenna
-                          design doesn&apos;t. Look before you let it run:
-                        </div>
-                        <ul className="design-trust-advisory">
-                          {d.advisory.map((a, i) => (
-                            <li key={i} className={`sev-${a.severity}`}>
-                              line {a.line}: {a.message}
-                            </li>
-                          ))}
-                        </ul>
-                      </>
-                    ) : (
-                      <div className="design-trust-advisory-head">
-                        Nothing unusual — it only builds antenna geometry.
-                      </div>
-                    )}
-                    <div className="design-trust-actions">
-                      <button
-                        className="design-trust-btn"
-                        disabled={isBusy}
-                        onClick={() => onTrust(d.name, false)}
-                      >
-                        Allow it to run
-                      </button>
-                      <button
-                        className="design-trust-btn is-edits"
-                        disabled={isBusy}
-                        onClick={() => onTrust(d.name, true)}
-                        title="For a design you're writing yourself — won't ask again when you save changes"
-                      >
-                        Allow + my edits
-                      </button>
-                    </div>
-                    <div className="design-trust-note">
-                      A design is a small program that runs on your computer.
-                      Only allow ones from people you trust.
-                    </div>
-                  </div>
-                )}
-              </li>
-            );
-          })}
-        </ul>
-      )}
-    </div>
-  );
-}
-
-// Searchable antenna picker: merges the old filter box + grouped <select> into
-// one combobox (type to filter, ▾ to open) so it fits on one line beside the
-// variant select. Keeps the family grouping the native <optgroup> list had.
-function GeometryCombobox({
-  groups,
-  selected,
-  currentLabel,
-  filter,
-  setFilter,
-  onSelect,
-}: {
-  groups: { fam: string; label: string; items: ExampleDescriptor[] }[];
-  selected: string;
-  currentLabel: string;
-  filter: string;
-  setFilter: (s: string) => void;
-  onSelect: (name: string) => void;
-}) {
-  const [open, setOpen] = useState(false);
-  const [active, setActive] = useState(0);
-  const rootRef = useRef<HTMLDivElement>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const flat = groups.flatMap((g) => g.items);
-
-  // Close (and clear the filter) on an outside click.
-  useEffect(() => {
-    if (!open) return;
-    const onDoc = (e: MouseEvent) => {
-      if (rootRef.current && !rootRef.current.contains(e.target as Node)) {
-        setOpen(false);
-        setFilter("");
-      }
-    };
-    document.addEventListener("mousedown", onDoc);
-    return () => document.removeEventListener("mousedown", onDoc);
-  }, [open, setFilter]);
-
-  const choose = (name: string) => {
-    onSelect(name);
-    setFilter("");
-    setOpen(false);
-    inputRef.current?.blur();
-  };
-
-  const onKey = (e: React.KeyboardEvent) => {
-    if (e.key === "ArrowDown") {
-      e.preventDefault();
-      setOpen(true);
-      setActive((i) => Math.min(flat.length - 1, i + 1));
-    } else if (e.key === "ArrowUp") {
-      e.preventDefault();
-      setActive((i) => Math.max(0, i - 1));
-    } else if (e.key === "Enter") {
-      if (open && flat[active]) {
-        e.preventDefault();
-        choose(flat[active].name);
-      }
-    } else if (e.key === "Escape") {
-      setOpen(false);
-      setFilter("");
-      inputRef.current?.blur();
-    }
-  };
-
-  return (
-    <div className="combobox" ref={rootRef}>
-      <input
-        ref={inputRef}
-        className="geometry-filter combobox-input"
-        type="text"
-        role="combobox"
-        aria-expanded={open}
-        aria-label="antenna"
-        placeholder="search antennas…"
-        value={open ? filter : currentLabel}
-        onChange={(e) => {
-          setFilter(e.target.value);
-          setOpen(true);
-          setActive(0);
-        }}
-        onFocus={() => setOpen(true)}
-        onKeyDown={onKey}
-      />
-      <span
-        className="combobox-caret"
-        aria-hidden="true"
-        onMouseDown={(e) => {
-          e.preventDefault();
-          if (open) {
-            setOpen(false);
-          } else {
-            inputRef.current?.focus();
-            setOpen(true);
-          }
-        }}
-      >
-        ▾
-      </span>
-      {open && (
-        <ul className="combobox-list" role="listbox">
-          {flat.length === 0 ? (
-            <li className="combobox-empty">no antennas match</li>
-          ) : (
-            groups.map((g) => (
-              <li key={g.fam} className="combobox-group">
-                <div className="combobox-group-label">{g.label}</div>
-                <ul>
-                  {g.items.map((ex) => {
-                    const idx = flat.indexOf(ex);
-                    return (
-                      <li
-                        key={ex.name}
-                        role="option"
-                        aria-selected={ex.name === selected}
-                        className={`combobox-option${
-                          ex.name === selected ? " is-selected" : ""
-                        }${idx === active ? " is-active" : ""}`}
-                        onMouseDown={(e) => {
-                          e.preventDefault();
-                          choose(ex.name);
-                        }}
-                        onMouseEnter={() => setActive(idx)}
-                      >
-                        {ex.label}
-                      </li>
-                    );
-                  })}
-                </ul>
-              </li>
-            ))
-          )}
-        </ul>
-      )}
-    </div>
-  );
-}
-
-// Band picker — a click-only dropdown, deliberately NOT a native <select>.
-// A focused <select> captures the arrow keys, which would fight the sticky
-// meas-freq dial (the "physical dial survives focus loss" affordance): the dial
-// shows armed but arrows would drive the pulldown. This is a plain <button> +
-// popover, so it never captures arrows — they always flow to the armed knob.
-function BandDropdown({
-  bands,
-  value,
-  onSelect,
-  disabled,
-  ariaLabel,
-}: {
-  bands: BandSpec[];
-  value: string;
-  onSelect: (key: string) => void;
-  disabled?: boolean;
-  ariaLabel: string;
-}) {
-  const [open, setOpen] = useState(false);
-  const rootRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!open) return;
-    const onDoc = (e: MouseEvent) => {
-      if (rootRef.current && !rootRef.current.contains(e.target as Node)) {
-        setOpen(false);
-      }
-    };
-    document.addEventListener("mousedown", onDoc);
-    return () => document.removeEventListener("mousedown", onDoc);
-  }, [open]);
-
-  const current = bands.find((b) => b.key === value) ?? bands[0];
-
-  return (
-    <div className="band-dropdown" ref={rootRef}>
-      <button
-        type="button"
-        className="band-select band-dropdown-trigger"
-        aria-haspopup="listbox"
-        aria-expanded={open}
-        aria-label={ariaLabel}
-        disabled={disabled}
-        onClick={() => setOpen((o) => !o)}
-      >
-        <span className="band-dropdown-value">{current?.label ?? ""}</span>
-        <span className="band-dropdown-caret" aria-hidden="true">
-          ▾
-        </span>
-      </button>
-      {open && !disabled && (
-        <ul className="band-dropdown-list" role="listbox" aria-label={ariaLabel}>
-          {bands.map((b) => (
-            <li
-              key={b.key}
-              role="option"
-              aria-selected={b.key === value}
-              className={`band-dropdown-option${
-                b.key === value ? " is-selected" : ""
-              }`}
-              onMouseDown={(e) => {
-                e.preventDefault();
-                onSelect(b.key);
-                setOpen(false);
-              }}
-            >
-              {b.label}
-            </li>
-          ))}
-        </ul>
-      )}
-    </div>
-  );
-}
-
-// Translate a knob's optional layout hint into inline grid-placement
-// styles. Returns undefined when nothing is set so auto-flow fields stay
-// untouched. `col_span` / `row_span` use the CSS `span N` form; an explicit
-// `col` / `row` pins the start line (1-indexed). `col` + `col_span`
-// together place a spanning field at a fixed column.
-function layoutStyle(layout?: KnobLayout | null): CSSProperties | undefined {
-  if (!layout) return undefined;
-  const style: CSSProperties = {};
-  const colStart = layout.col ?? null;
-  const rowStart = layout.row ?? null;
-  const colSpan = layout.col_span ?? null;
-  const rowSpan = layout.row_span ?? null;
-  if (colStart != null) {
-    style.gridColumn = colSpan != null ? `${colStart} / span ${colSpan}` : `${colStart}`;
-  } else if (colSpan != null) {
-    style.gridColumn = `span ${colSpan}`;
-  }
-  if (rowStart != null) {
-    style.gridRow = rowSpan != null ? `${rowStart} / span ${rowSpan}` : `${rowStart}`;
-  } else if (rowSpan != null) {
-    style.gridRow = `span ${rowSpan}`;
-  }
-  return Object.keys(style).length ? style : undefined;
-}
-
-function ParamForm({
-  schema,
-  values,
-  onChange,
-  pathPrefix = [],
-  disabledFields,
-  opt,
-}: {
-  schema: SchemaItem[];
-  values: ParamValueBag;
-  onChange: (path: (string | number)[], value: number | string | boolean) => void;
-  pathPrefix?: (string | number)[];
-  // Param names that should render as disabled even though they're
-  // visible in the schema — e.g. to grey out a control whose effect
-  // depends on the active backend. Currently unused (kept as a general
-  // mechanism); daisy_chain used to rely on it before build_network()
-  // made the single-feed hexbeam engine-agnostic.
-  disabledFields?: Set<string>;
-  // Optimiser integration (top-level rail only). `settings` overrides a knob's
-  // effective min/max/step; `onContext` opens that knob's right-click menu;
-  // `onToggleVary` flips a knob's "Optimize this knob" flag (the `o` shortcut,
-  // parallel to the menu checkbox).
-  opt?: {
-    settings: Record<string, KnobOpt>;
-    onContext: (name: string, e: React.MouseEvent) => void;
-    onToggleVary: (name: string) => void;
-  };
-}) {
-  return (
-    <>
-      {schema.map((item) => {
-        if (isGroup(item)) {
-          const countRaw = values[item.repeat_count];
-          const count = typeof countRaw === "number" ? Math.round(countRaw) : 0;
-          const instances = values[item.name];
-          if (!Array.isArray(instances)) return null;
-          return (
-            <div key={item.name} className="param-group">
-              {Array.from({ length: Math.min(count, instances.length) }, (_, i) => (
-                <div key={`${item.name}-${i}`} className="param-group-instance">
-                  <div className="param-group-header">
-                    {item.label_template.replace("{i}", String(i))}
-                  </div>
-                  {/* Wrap the band's controls in their own .param-grid so they
-                      pack 3-across just like the top-level rail. Without this
-                      the nested ParamForm block-stacks one control per row. */}
-                  <div className="param-grid is-knobs">
-                    <ParamForm
-                      schema={item.params}
-                      values={instances[i]}
-                      onChange={onChange}
-                      pathPrefix={[...pathPrefix, item.name, i]}
-                      disabledFields={disabledFields}
-                    />
-                  </div>
-                </div>
-              ))}
-            </div>
-          );
-        }
-        // Scalar leaf.
-        if (!applyVisibility(item, values)) return null;
-        const currentRaw = values[item.name];
-        const currentNum =
-          typeof currentRaw === "number" ? currentRaw : Number(item.default);
-        const currentStr =
-          typeof currentRaw === "string" ? currentRaw : String(item.default);
-
-        // Resolve dynamic min/max from a sibling enum's currently-
-        // selected option, when configured. Falls back to the static
-        // min/max if the lookup misses.
-        let effMin = item.min ?? 0;
-        let effMax = item.max ?? 1;
-        const rfe = item.range_from_enum_option;
-        if (rfe) {
-          const siblingVal = values[rfe.param];
-          const siblingSchema = schema.find(
-            (s) => !isGroup(s) && s.name === rfe.param,
-          ) as SchemaParamSpec | undefined;
-          const opts = siblingSchema?.enum_options;
-          if (opts && typeof siblingVal === "string") {
-            const opt = opts.find((o) => o.value === siblingVal);
-            if (opt) {
-              const lo = opt[rfe.min_key];
-              const hi = opt[rfe.max_key];
-              if (typeof lo === "number") effMin = lo;
-              if (typeof hi === "number") effMax = hi;
-            }
-          }
-        }
-
-        if (item.kind === "bool") {
-          const checked = Boolean(currentRaw ?? item.default);
-          const isDisabled = disabledFields?.has(item.name) ?? false;
-          return (
-            <div key={item.name} className="field field-bool" style={layoutStyle(item.layout)}>
-              <label
-                className={`field-bool-label${isDisabled ? " field-disabled" : ""}`}
-              >
-                <input
-                  type="checkbox"
-                  checked={checked}
-                  disabled={isDisabled}
-                  onChange={(e) =>
-                    onChange(
-                      [...pathPrefix, item.name],
-                      (e.target as HTMLInputElement).checked,
-                    )
-                  }
-                />
-                <span>{item.label}</span>
-              </label>
-            </div>
-          );
-        }
-
-        if (item.kind === "enum") {
-          const opts = item.enum_options ?? [];
-          return (
-            <div key={item.name} className="field field-enum" style={layoutStyle(item.layout)}>
-              <label>
-                <span>{item.label}</span>
-              </label>
-              <select
-                value={currentStr}
-                onChange={(e) => {
-                  const next = (e.target as HTMLSelectElement).value;
-                  onChange([...pathPrefix, item.name], next);
-                  // On-change side effect: set a sibling's value to a
-                  // key from the new enum option. Fan_dipole uses this
-                  // to snap freq to the band's default.
-                  const oc = item.on_change_set;
-                  if (oc) {
-                    const opt = opts.find((o) => o.value === next);
-                    if (opt) {
-                      const k = opt[oc.from_enum_key];
-                      if (typeof k === "number" || typeof k === "string") {
-                        onChange([...pathPrefix, oc.set], k);
-                      }
-                    }
-                  }
-                }}
-              >
-                {opts.map((o) => (
-                  <option key={o.value} value={o.value}>{o.label}</option>
-                ))}
-              </select>
-            </div>
-          );
-        }
-
-        const shown =
-          item.kind === "int"
-            ? String(Math.round(currentNum))
-            : currentNum.toFixed(item.precision);
-        // Every float/int param is a rotary knob: label on top, dial in the
-        // middle, value on the bottom. (The slider alternative and its toggle
-        // were retired — knobs are the brand.)
-        // Per-knob optimiser override: display extents + manual step come from
-        // the knob's menu when set, and `vary` marks it a free variable.
-        const ko = opt?.settings[item.name];
-        const knobMin = ko ? ko.dispMin : effMin;
-        const knobMax = ko ? ko.dispMax : effMax;
-        const knobStep = ko?.step ?? item.step ?? 0.001;
-        return (
-          <div
-            key={item.name}
-            className={`field field-knob${ko?.vary ? " is-opt-var" : ""}`}
-            style={layoutStyle(item.layout)}
-            onContextMenu={opt ? (e) => opt.onContext(item.name, e) : undefined}
-            // `o` toggles this knob's "Optimize this knob" flag while it's
-            // focused — the keyboard parallel to the right-click menu. The event
-            // bubbles up from the focused role="slider" Knob; the edit <input>
-            // stops propagation, so typing a value never triggers it. Ignore it
-            // when a modifier is held (reserved for other shortcuts) or on
-            // auto-repeat (holding the key mustn't flip-flop the flag).
-            onKeyDown={
-              opt
-                ? (e) => {
-                    if (
-                      e.key.toLowerCase() === "o" &&
-                      !e.ctrlKey &&
-                      !e.metaKey &&
-                      !e.altKey &&
-                      !e.repeat
-                    ) {
-                      e.preventDefault();
-                      opt.onToggleVary(item.name);
-                    }
-                  }
-                : undefined
-            }
-          >
-            <span
-              className="knob-label"
-              title={item.name === item.label ? item.label : `${item.label} · param: ${item.name}`}
-            >
-              {item.label}
-            </span>
-            <Knob
-              knobId={[...pathPrefix, item.name].join(".")}
-              value={currentNum}
-              min={knobMin}
-              max={knobMax}
-              step={knobStep}
-              precision={item.kind === "int" ? 0 : item.precision}
-              unit={item.unit}
-              label={item.label}
-              onChange={(v) => onChange([...pathPrefix, item.name], v)}
-            />
-            <span className="knob-value">{shown}{item.unit ?? ""}</span>
-          </div>
-        );
-      })}
-    </>
-  );
-}
 
 // label_template substitutions for ResultGroupItem:
 //   {i}            → 0-based index
@@ -1655,7 +806,7 @@ const OPT_OBJECTIVES: OptObjective[] = ["swr", "resonance"];
 // search; display extents are the knob's own slider range; step is the manual
 // turn granularity (the optimiser itself is continuous). Absent for a knob =
 // schema defaults.
-type KnobOpt = {
+export type KnobOpt = {
   vary: boolean;
   optMin: number;
   optMax: number;
